+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/ClaimValue.java
@@ -4,6 +4,7 @@ import gov.cms.bfd.server.ng.converter.NonZeroIntConverter;
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Embeddable;
+import java.math.BigDecimal;
 import java.util.Optional;
 import lombok.Getter;
 
@@ -18,20 +19,20 @@ class ClaimValue {
   private Optional<String> claimValueCode;
 
   @Column(name = "clm_val_amt")
-  private double claimValueAmount;
+  private BigDecimal claimValueAmount;
 
   private static final String VALUE_CODE_DISPROPORTIONATE = "18";
   private static final String VALUE_CODE_IME = "19";
 
-  Optional<Double> getDisproportionateAmount() {
+  Optional<BigDecimal> getDisproportionateAmount() {
     return getAmountForCode(VALUE_CODE_DISPROPORTIONATE);
   }
 
-  Optional<Double> getImeAmount() {
+  Optional<BigDecimal> getImeAmount() {
     return getAmountForCode(VALUE_CODE_IME);
   }
 
-  private Optional<Double> getAmountForCode(String code) {
+  private Optional<BigDecimal> getAmountForCode(String code) {
     return claimValueCode.flatMap(
         c -> {
           if (c.equals(code)) {
